+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- New `addStickerPack` command
+
 ## [0.12.0] - 2023-08-11
 **Attention**: Now requires native libsignal-client version 0.30.0